server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pdfplumber==0.10.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
#pip install -r requirements.txt
